Automatisierte Git-Tests

+++ src/main/java/htwberlin/webtech/ss24/fitnessplaner/controller/ExerciseRestController.java
@@ -1,14 +1,10 @@
-package htwberlin.webtech.ss24.fitnessplaner.web;
+package htwberlin.webtech.ss24.fitnessplaner.controller;
 
 import htwberlin.webtech.ss24.fitnessplaner.model.Exercise;
 import htwberlin.webtech.ss24.fitnessplaner.web.persistence.ExerciseManipulationRequest;
 import htwberlin.webtech.ss24.fitnessplaner.web.service.ExerciseService;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -16,8 +12,6 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.logging.Logger;
 import java.util.logging.Level;
-
-
 
 @RestController
 @RequestMapping("/workoutplan")
@@ -32,6 +26,9 @@ public class ExerciseRestController {
 
     @GetMapping("/search")
     public ResponseEntity<List<Exercise>> searchExercise(@AuthenticationPrincipal OidcUser user, Model model, @RequestParam String query) {
+        if (user == null) {
+            return ResponseEntity.status(401).build();
+        }
         List<Exercise> exercises = exerciseService.searchByName(query);
         if (exercises.isEmpty()) {
             return ResponseEntity.notFound().build();
@@ -41,6 +38,9 @@ public class ExerciseRestController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Exercise> getExerciseById(@AuthenticationPrincipal OidcUser user, Model model, @PathVariable Long id) {
+        if (user == null) {
+            return ResponseEntity.status(401).build();
+        }
         Exercise exercise = exerciseService.findById(id);
         if (exercise == null) {
             return ResponseEntity.notFound().build();
@@ -48,22 +48,11 @@ public class ExerciseRestController {
         return ResponseEntity.ok(exercise);
     }
 
-
-    @GetMapping("/history")
-    public ResponseEntity<List<Exercise>> getAllExercises(@AuthenticationPrincipal OidcUser user) {
-        if (user != null) {
-            String email = user.getEmail();
-            List<Exercise> exercises = exerciseService.findByOwner(email);
-            return ResponseEntity.ok(exercises);
-        } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-    }
-
-
-
     @PostMapping
     public ResponseEntity<Exercise> createExercise(@AuthenticationPrincipal OidcUser user, Model model, @RequestBody ExerciseManipulationRequest request) {
+        if (user == null) {
+            return ResponseEntity.status(401).build();
+        }
         logger.log(Level.INFO, "Received request to create exercise: {0}", request);
         if (request.getOwner() == null || request.getOwner().isEmpty()) {
             logger.log(Level.SEVERE, "Error: Owner darf nicht null oder leer sein.");
@@ -79,12 +68,11 @@ public class ExerciseRestController {
         }
     }
 
-
-
-
-
     @PutMapping("/{id}")
     public ResponseEntity<Exercise> updateExercise(@AuthenticationPrincipal OidcUser user, Model model, @PathVariable Long id, @RequestBody ExerciseManipulationRequest request) {
+        if (user == null) {
+            return ResponseEntity.status(401).build();
+        }
         Exercise exercise = exerciseService.update(id, request);
         if (exercise == null) {
             return ResponseEntity.notFound().build();
@@ -94,6 +82,9 @@ public class ExerciseRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteExercise(@AuthenticationPrincipal OidcUser user, Model model, @PathVariable Long id) {
+        if (user == null) {
+            return ResponseEntity.status(401).build();
+        }
         boolean success = exerciseService.deleteById(id);
         if (!success) {
             return ResponseEntity.notFound().build();
